refactor: move ucsc-tracks-api into @repo/shared (#1539)
