+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/components/TracksSelector/components/TracksSelectionPanel/types.ts
@@ -1,4 +1,4 @@
-import { UcscTrackNode } from "@/utils/ucsc-tracks-api/entities";
+import { UcscTrackNode } from "@repo/shared/utils/ucsc-tracks-api/types";
 import { Table } from "@tanstack/react-table";
 
 export interface Props {
