+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/components/TracksSelector/hooks/UseTable/columnDef.ts
@@ -1,7 +1,7 @@
-import { UcscTrackNode } from "@/utils/ucsc-tracks-api/entities";
 import { LABEL } from "@databiosphere/findable-ui/lib/apis/azul/common/entities";
 import { COLUMN_IDENTIFIER } from "@databiosphere/findable-ui/lib/components/Table/common/columnIdentifier";
 import { RowSelectOrExpand } from "@repo/shared/components/Table/components/TableCell/components/RowSelectOrExpand/rowSelectOrExpand";
+import { UcscTrackNode } from "@repo/shared/utils/ucsc-tracks-api/types";
 import { ColumnDef } from "@tanstack/react-table";
 import { CATEGORY_CONFIGS } from "./categoryConfigs";
