+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/components/TracksSelector/hooks/UseTable/types.ts
@@ -1,4 +1,4 @@
-import { UcscTrack } from "@/utils/ucsc-tracks-api/entities";
+import { UcscTrack } from "@repo/shared/utils/ucsc-tracks-api/types";
 
 export type Track = Omit<UcscTrack, "isComposite"> & {
   groupId: string;
