+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/types.ts
@@ -1,5 +1,5 @@
 import { StepProps } from "@/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/types";
-import { UcscTrackNode } from "@/utils/ucsc-tracks-api/entities";
+import { UcscTrackNode } from "@repo/shared/utils/ucsc-tracks-api/types";
 import { Table } from "@tanstack/react-table";
 
 export interface Props extends Pick<StepProps, "onConfigure" | "stepKey"> {
