+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/hooks/UseUCSCTracks/types.ts
@@ -1,4 +1,4 @@
-import { UcscTrackNode } from "@/utils/ucsc-tracks-api/entities";
+import { UcscTrackNode } from "@repo/shared/utils/ucsc-tracks-api/types";
 
 export interface UseUCSCTracks {
   data?: UcscTrackNode[];
