+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/hooks/UseLaunchGalaxy/types.ts
@@ -1,12 +1,12 @@
 import { COLUMN_TYPE } from "@/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/SampleSheetClassificationStep/types";
 import { Strandedness } from "@/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/StrandednessStep/types";
 import { EnaSequencingReads } from "@/utils/galaxy-api/entities";
-import { UcscTrack } from "@/utils/ucsc-tracks-api/entities";
 import {
   ConfiguredInput,
   PrimaryContrasts,
 } from "@/views/WorkflowInputsView/hooks/UseConfigureInputs/types";
 import type { Workflow } from "@repo/shared/apis/workflow";
+import { UcscTrack } from "@repo/shared/utils/ucsc-tracks-api/types";
 
 // Base configured values shared across all scopes
 export interface BaseConfiguredValue {
